swap the precedence of default_context and builder
since the default_context would hold a potentially lazy build object and
would later be used once created, it would never reach the coderef in
builder. But that is what would be preferable.

swapping around a lot made it such that if there is a coderef for the
builder, then that will be used, otherwise a one-off lazely build object
stored in default_context.

surely, the default will croak.